requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.4
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3
flake8>=7.0
black>=24.0
